+++ package.json
@@ -20,6 +20,7 @@
     "react-scripts": "4.0.3",
     "styled-components": "^5.3.0",
     "typescript": "^4.1.2",
+    "uuid": "^8.3.2",
     "web-vitals": "^1.0.1"
   },
   "scripts": {
@@ -48,6 +49,7 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.172",
-    "@types/styled-components": "^5.1.11"
+    "@types/styled-components": "^5.1.11",
+    "@types/uuid": "^8.3.4"
   }
 }
